+++ getl/common/hive_table.py
@@ -16,7 +16,7 @@ class HiveTable:
     database_name: str
     table_name: str
 
-    def create(self, location: str, db_schema: str = "") -> None:
+    def create(self, location: str, db_schema: str = "", columns: list = None) -> None:
         """Create hive table."""
         LOGGER.info('Create Hive table: "%s.%s"', self.database_name, self.table_name)
         create_table = f"CREATE TABLE IF NOT EXISTS {self.table_name}"
@@ -27,9 +27,11 @@ class HiveTable:
 
         self.spark.sql(f"CREATE DATABASE IF NOT EXISTS {self.database_name}")
         self.spark.sql(f"USE {self.database_name}")
-        self.spark.sql(
-            f"""
-            {create_table}
-            USING DELTA LOCATION "{location}"
-        """
-        )
+        if columns:
+            columns_as_string = ",".join(columns)
+            self.spark.sql(
+                f"{create_table} USING DELTA PARTITIONED BY "
+                f"({columns_as_string}) LOCATION {location}"
+            )
+        else:
+            self.spark.sql(f"{create_table} USING DELTA LOCATION {location}")